Update world interface location and VenuePartygoers display
