+++ src/components/molecules/VenuePartygoers/VenuePartygoers.scss
@@ -1,5 +1,7 @@
-.venue-partygoers-container {
-  font-size: 0.8rem;
-  opacity: 0.7;
-  margin-left: 40px;
+@import "scss/constants";
+
+.VenuePartygoers {
+  color: var(--content--over-50a);
+  font-size: font-size--root(0.8);
+  margin-left: font-size--root(2.5);
 }
