+++ src/components/organisms/WorldShowcase/WorldShowcase.scss
@@ -23,7 +23,7 @@ $content-margin--left: font-size--root(3);
   &__nav {
     width: 100%;
     height: $nav-height;
-    background-color: var(--content--under-24a);
+    background-color: var(--content--under-25a);
     display: grid;
     place-items: center start;
   }
