+++ src/components/templates/PosterHall/components/PosterPreview/PosterPreview.scss
@@ -2,7 +2,7 @@
 
 $poster-preview-height: 300px;
 $poster-preview-width: 300px;
-$active-poster-shadows: 0 10px 20px 0 var(--primary--under-56a);
+$active-poster-shadows: 0 10px 20px 0 var(--primary--under-50a);
 
 .PosterPreview {
   display: flex;
