+++ src/hooks/useCurrentWorld.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { useFirestore, useFirestoreDocData } from "reactfire";
 
-import { World } from "api/admin";
+import { World } from "api/world";
 
 import { ReactHook } from "types/utility";
