+++ src/hooks/useWorldEdit.ts
@@ -1,6 +1,6 @@
 import { useFirestore, useFirestoreDocData } from "reactfire";
 
-import { World } from "api/admin";
+import { World } from "api/world";
 
 import { worldConverter } from "utils/converters";
 import { WithId } from "utils/id";
