+++ src/scss/colors.scss
@@ -11,8 +11,8 @@ $colors-map: (
   danger--under: #bc271a
 );
 
-$shades-enum: (5, 10, 14, 20, 50);
-$alphas-enum: (0.1, 0.18, 0.2, 0.24, 0.35, 0.4, 0.5, 0.56, 0.65, 0.8);
+$shades-enum: (5, 10, 15, 20, 50);
+$alphas-enum: (0.1, 0.2, 0.25, 0.3, 0.4, 0.5, 0.6, 0.7, 0.75, 0.8, 0.9);
 
 @function rgb-triplet($hex) {
   @return "" + red($hex) + "," + green($hex) + "," + blue($hex);
