+++ src/scss/constants.scss
@@ -7,7 +7,7 @@ $content--over: var(--content--over);
 $content--under: var(--content--under);
 
 // @debt these should probably be replaced with var(--primary-...) referencing theme colors
-$primary--lightest: #a179ff; // replace with var(--primary--under-lighter-14pp) // just an approximation from the available
+$primary--lightest: #a179ff; // replace with var(--primary--under-lighter-15pp) // just an approximation from the available
 $primary--light: #9669ff; // replace with var(--primary--under-lighter-10pp) // just an approximation from the available
 $primary--live-hover: #8755fb; // replace with var(--primary--under-lighter-5pp) // just an approximation from the available
 $primary--darker: #3f247d; // replace with var(--primary--under-darker-20pp) // just an approximation from the available
