+++ src/types/Firestore.ts
@@ -1,6 +1,6 @@
 import { Settings } from "./settings";
 
-import { World } from "api/admin";
+import { World } from "api/world";
 
 import { AuditoriumSeatedUser } from "types/auditorium";
 import {
@@ -25,6 +25,7 @@ export interface Experience {
   reactions: Record<string, Reaction>;
   tables: Record<string, Record<string, Table>>;
 }
+
 export interface UserVisit {
   timeSpent: number;
 }
@@ -114,6 +115,7 @@ export interface FirestoreOrdered {
 export interface DistributedCounterValue {
   value: number;
 }
+
 export interface DistributedCounterShard {
   count: number;
 }
